python/requirements.txt
paho-mqtt
asyncio-mqtt
aiofiles
requests
numpy
scipy
matplotlib
smbus2
spidev
RPi.GPIO
wiringpi
configparser